I updated the classes to use the new API for pulling repository information, and started tracking watchers and forks for each repo.  On the main repo list pages, I'm now showing that info instead of the last commit, since it was confusing to have nothing there until you fetched the commits.  I'm thinking of adding last_updated, but I'll have to add that to the api data, since it doesn't seem to be there.
Since it was confusing to have the repo cells switch data whether you
had looked at the commits or not, I turned off showing the last_commit
data at all.

I also changed the functionality to only post for data when it's needed,
rather than on app load, so people can use the app without actually
having a GitHub account typed in.  If so, the first block of menu items
does not show up - they're all meant to be used with an authenticated
account.

I also added some extra menu options that I would like to implement in
the future.